Many changes. Finally porting them over from the main Python tool.
Includes mode selection. Modes now determine which controls are active.

Delayed save now available.

Update checking and season URL grabbing is now done as the application
runs. This ensures that the user does not need to update solely for the
correct live page URL, along with ensuring that they have the latest
version.